EUI-2844: Wrapping error (#88)
* Update due-date.component.scss

The Due Date component no longer allows the text to wrap. Wrapping causes weird visuals and should be prevented in all forseeable instances. If we encounter an issue where the text is simply too long and SHOULD be wrapped in the future, we'd need to take a different approach with the border.

* Versioning

Upping the package version numbers.

Co-authored-by: Paul Graham <paul@smartcasual.io>