Preselection and options updates
Added functionality around changing the options provided and also around changing the preselection.

Changing the preselection now acts as changing the selection UNLESS the user has made an active selection, after which it is ignored.

Changing the options now makes sure to remove any selected items that are no longer in the list of available options.

Whenever the selection is changed, an event is appropriately emitted.