EUI-2910: Checkbox list (#91)
* Checkbox List component

Initial commit of the Checkbox List component. There are presently no unit tests in place for it but there is a test harness to observe and test the behaviour of the component.

* Linting

Fixed a bunch of linting errors.

* Preselection and options updates

Added functionality around changing the options provided and also around changing the preselection.

Changing the preselection now acts as changing the selection UNLESS the user has made an active selection, after which it is ignored.

Changing the options now makes sure to remove any selected items that are no longer in the list of available options.

Whenever the selection is changed, an event is appropriately emitted.

* Changing label function

Added a mechanism in the test harness to demonstrate that the label function can be changed and that this automatically results in the labels changing in the component.

* Matching local reposiotry

* Unit tests for checkbox-list

* Update checkbox-list.component.spec.ts

* Checkbox List component

Initial commit of the Checkbox List component. There are presently no unit tests in place for it but there is a test harness to observe and test the behaviour of the component.

* Linting

Fixed a bunch of linting errors.

* Preselection and options updates

Added functionality around changing the options provided and also around changing the preselection.

Changing the preselection now acts as changing the selection UNLESS the user has made an active selection, after which it is ignored.

Changing the options now makes sure to remove any selected items that are no longer in the list of available options.

Whenever the selection is changed, an event is appropriately emitted.

* Changing label function

Added a mechanism in the test harness to demonstrate that the label function can be changed and that this automatically results in the labels changing in the component.

* Added more unit tests

* Update checkbox-list.component.spec.ts

Added a test to make sure that the checkboxes are removed if the options are removed.

* Version

Upping version numbers.

* Add last unit tests

Also edited one comment's spelling

* Fixing lint errors

* Line feeds

Lint seems to have a thing for ending files with a line feed... so I've done that for the HTML and SCSS files too.

* Update app.component.ts

Lint error for properties appearing after functions.

* Versions

Up to version 0.1.47, since 0.1.46 was taken for something else.

Co-authored-by: Paul Graham <paul@smartcasual.io>
Co-authored-by: connorpgpmcelroy <connor.mcelroy@hmcts.net>